+++ src/features/book-studio/index.ts
@@ -1,4 +1,6 @@
 export * from './constants';
 export * from './api/storybookApi';
 export * from './helpers/storyDraft';
+export * from './helpers/storyMapping';
+export * from './hooks/useStoryAiController';
 export type * from './types';